+++ src/main/java/net/sodium/client/gl/device/GLRenderDevice.java
@@ -214,7 +214,7 @@ public class GLRenderDevice implements RenderDevice {
             this.bindBuffer(GlBufferTarget.ARRAY_BUFFER, buffer);
 
             CommandContext ctx = VulkanicAPI.getImmediateContext();
-            ByteBuffer buf = VulkanicAPI.mapBuffer(ctx, GlBufferTarget.ARRAY_BUFFER.getTargetParameter(), (int)offset, (int)length, flags.getBitField());
+            ByteBuffer buf = VulkanicAPI.mapBuffer(ctx, GlBufferTarget.ARRAY_BUFFER.getTargetParameter(), offset, length, flags.getBitField());
 
             if (buf == null) {
                 throw new RuntimeException("Failed to map buffer");
